feat(#39): remove system logger
