+++ src/wazuh_qa_framework/generic_modules/logging/base_logger.py
@@ -15,6 +15,7 @@ This module contains the following:
 """
 import logging
 import sys
+import inspect
 import os
 
 from wazuh_qa_framework.generic_modules.exceptions.exceptions import ValidationError
@@ -48,11 +49,11 @@ LOG_COLORS = {
 }
 
 FORMATERS = {
-    'basic': logging.Formatter('%(asctime)s — %(levelname)s — %(message)s')
+    'basic': logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
 }
 
 
-class BaseLogger:
+class BaseLogger():
     """Logger class to manage modules logging.
 
     This class defines the main behavior of the loggers and will be used as a base class for the rest of the loggers
@@ -66,6 +67,7 @@ class BaseLogger:
         handlers (list(str)): Logging handlers.
         logging_file (str): File path were save the logging if the file handler has been specified.
         output_color (boolean): True for logging with output colors, False otherwise.
+        output_source (boolean): True for logging with the source of the message, False otherwise.
 
     Attributes:
         name (str): Logger name.
@@ -75,9 +77,10 @@ class BaseLogger:
         handlers (list(str)): Logging handlers.
         logging_file (str): File path were save the logging if the file handler has been specified.
         output_color (boolean): True for logging with output colors, False otherwise.
+        output_source (boolean): True for logging with the source of the message, False otherwise.
     """
-    def __init__(self, name, level='info', formatter='basic', handlers=None, logging_file=None,
-                 output_color=True):
+    def __init__(self, name, level='info', formatter='basic', handlers=None, logging_file=None, output_color=True,
+                 output_source=False):
         self.name = name
         self.__logger = logging.getLogger(name)
         self.level = level
@@ -85,6 +88,7 @@ class BaseLogger:
         self.logging_file = logging_file
         self.handlers = handlers
         self.output_color = output_color
+        self.output_source = output_source
 
     @property
     def name(self):
@@ -160,7 +164,6 @@ class BaseLogger:
         if new_formatter in FORMATERS.keys():
             self.__formatter = FORMATERS[new_formatter]
         else:
-            self.__formatter = FORMATERS['basic']
             raise ValidationError(f"{new_formatter} is not an accepted formatter. Expected one of"
                                   f" {list(FORMATERS.keys())}")
 
@@ -219,7 +222,7 @@ class BaseLogger:
         handlers = new_handlers if new_handlers is not None else ['sys_output']
 
         if type(handlers) != list:
-            ValidationError(f"It was expected to receive a list of allowed handlers: f{HANDLERS}")
+            raise ValidationError(f"It was expected to receive a list of allowed handlers: f{HANDLERS}")
 
         # Set logger handlers
         for handler in handlers:
@@ -243,7 +246,7 @@ class BaseLogger:
                     stream_handler.setFormatter(self.__formatter)
                     stream_handler.setLevel(self.__level)
                     self.__logger.addHandler(stream_handler)
-                    ValidationError(f"{handler} handler is not allowed")
+                    raise ValidationError(f"{handler} handler is not allowed")
 
         self.__handlers = handlers
 
@@ -256,6 +259,24 @@ class BaseLogger:
         """
         return self.__output_color
 
+    @property
+    def output_source(self):
+        """Getter to obtain output_source attribute.
+
+        Returns:
+            boolean: True if output source is enabled, False otherwise.
+        """
+        return self.__output_source
+
+    @output_source.setter
+    def output_source(self, new_output_source):
+        """Setter of output_source attribute.
+
+        Args:
+            new_output_source (boolean): New output source value.
+        """
+        self.__output_source = new_output_source
+
     @output_color.setter
     def output_color(self, new_output_color):
         """Setter of output_color attribute.
@@ -264,6 +285,32 @@ class BaseLogger:
             new_output_color (boolean): New output color value.
         """
         self.__output_color = new_output_color
+
+    def generate_logging_message(self, message, level='info'):
+        final_message = message
+        if self.output_source:
+            try:
+                curframe = inspect.currentframe()
+                calframe = inspect.getouterframes(curframe, 2)
+                calframe_origin = calframe[2]
+
+                origin_class = None
+                if 'self' in calframe_origin[0].f_locals:
+                    origin_class = calframe_origin[0].f_locals['self'].__class__.__name__
+
+                method = calframe_origin[3]
+                line = calframe_origin[2]
+                message = message + f" [{origin_class} {method}:{line}]" if origin_class \
+                    else message + f"[{method}:{line}]"
+
+            except Exception:
+                self.logger.warning("Error getting the source of the message")
+
+        if self.output_color:
+            color = LOG_COLORS[level]
+            final_message = f"{color}{message}{COLORS['CLEAR']}"
+
+        return final_message
 
     def log(self, message, level='info'):
         """Log a specific level message.
@@ -296,8 +343,7 @@ class BaseLogger:
         Args:
             message (str): Logging message.
         """
-        custom_message = f"{LOG_COLORS['debug']}{message}{COLORS['CLEAR']}" if self.output_color else message
-        self.logger.debug(custom_message)
+        self.logger.debug(self.generate_logging_message(message, level='debug'))
 
     def info(self, message):
         """INFO logging.
@@ -305,8 +351,7 @@ class BaseLogger:
         Args:
             message (str): Logging message.
         """
-        custom_message = f"{LOG_COLORS['info']}{message}{COLORS['CLEAR']}" if self.output_color else message
-        self.logger.info(custom_message)
+        self.logger.info(self.generate_logging_message(message, level='info'))
 
     def warning(self, message):
         """WARNING logging.
@@ -314,8 +359,7 @@ class BaseLogger:
         Args:
             message (str): Logging message.
         """
-        custom_message = f"{LOG_COLORS['warning']}{message}{COLORS['CLEAR']}" if self.output_color else message
-        self.logger.warning(custom_message)
+        self.logger.warning(self.generate_logging_message(message, level='warning'))
 
     def error(self, message):
         """ERROR logging.
@@ -323,8 +367,7 @@ class BaseLogger:
         Args:
             message (str): Logging message.
         """
-        custom_message = f"{LOG_COLORS['error']}{message}{COLORS['CLEAR']}" if self.output_color else message
-        self.logger.error(custom_message)
+        self.logger.error(self.generate_logging_message(message, level='error'))
 
     def critical(self, message):
         """CRITICAL logging.
@@ -332,8 +375,7 @@ class BaseLogger:
         Args:
             message (str): Logging message.
         """
-        custom_message = f"{LOG_COLORS['critical']}{message}{COLORS['CLEAR']}" if self.output_color else message
-        self.logger.critical(custom_message)
+        self.logger.critical(self.generate_logging_message(message, level='critical'))
 
     def __str__(self):
         """Redefine the logger object representation.
